% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_trees.R
\name{assign_trees}
\alias{assign_trees}
\title{Find matches between predicted and ground truth tree based on maximum area of overlap.}
\usage{
assign_trees(ground_truth, predictions)
}
\arguments{
\item{ground_truth}{A ground truth polygon in SpatialPolygonsDataFrame}

\item{prediction}{prediction polygons in SpatialPolygonsDataFrame}
}
\value{
A data frame with the crown ID matched to the prediction ID.
}
\description{
\code{assign_trees} implements the hungarian algorithm in \code{\link[clue]{solve_LSAP}} to match sets of trees
}
\examples{

}
