% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_jaccard.R
\name{calc_jaccard}
\alias{calc_jaccard}
\title{Collect jaccard statistic for ground truth and prediction polygons}
\usage{
calc_jaccard(assignment, ground_truth, predictions)
}
\arguments{
\item{assignment}{A dataframe that connects ground truth and prediction poylgons see \code{\link{assign_trees}} SpatialPolygonDataFrame of length 1}

\item{ground_truth}{A SpatialPolygonDataFrame of ground truth polygons}

\item{prediction}{A SpatialPolygonDataFrame of ground truth polygons}
}
\value{
dataframe of the jaccard overlap among polygon pairs
}
\description{
\code{calc_jaccard} finds the jaccard statistic for two input polygons, defined as the area of intersection over the area of union.
}
