% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_precision_recall.R
\name{compute_precision_recall}
\alias{compute_precision_recall}
\title{Compute precision and recall statistics between predicted tree boxes and ground truth data}
\usage{
compute_precision_recall(ground_truth, predictions, threshold = 0.5)
}
\arguments{
\item{ground_truth}{SpatialPolygonDataFrame of ground truth polygons}

\item{predictions}{SpatialPolygonDataFrame of prediction polygons}

\item{threshold}{Intersection over Union threshold. default is 0.5}
}
\value{
recall and precision scores for the plot
}
\description{
Compute precision and recall statistics between predicted tree boxes and ground truth data
}
\details{
The numeric identity of the crown is stored in a column named crown_ID. An error will be raised if this column does not exist.
}
