% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_plot.R
\name{evaluate_plot}
\alias{evaluate_plot}
\title{Compute evaluation statistics for one plot}
\usage{
evaluate_plot(submission, show = TRUE)
}
\arguments{
\item{submission}{A five column dataframe in the order plot_name, xmin, xmax, ymin, ymax. Each row is a predicted bounding box.}
}
\value{
recall and precision scores for the plot
}
\description{
Compute evaluation statistics for one plot
}
