% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_coordinates.R
\name{extract_coordinates}
\alias{extract_coordinates}
\title{Extract bounding box coordinates from SpatialPolygons object}
\usage{
extract_coordinates(x)
}
\arguments{
\item{fil}{"Character" filename of the .xml file}
}
\value{
a dataframe of tree annotations in the format xmin, xmax, ymin, ymax
}
\description{
Extract bounding box coordinates from SpatialPolygons object
}
