% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ground_truth.R
\name{load_ground_truth}
\alias{load_ground_truth}
\title{Load and overlay ground truth annotations for single plot evaluation}
\usage{
load_ground_truth(plot_name, show = TRUE)
}
\arguments{
\item{plot_name}{The name of plot as given by the filename (e.g "SJER_021.tif" -> SJER_021)}

\item{show}{Logical. Whether to show a plot of the ground truth data overlayed on the RGB image}
}
\value{
SpatialPolygonsDataFrame of ground truth boxes
}
\description{
Load and overlay ground truth annotations for single plot evaluation
}
