% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon_overlap_all.R
\name{polygon_overlap_all}
\alias{polygon_overlap_all}
\title{Find area of overlap among all sets of polygons}
\usage{
polygon_overlap_all(ground_truth, predictions)
}
\arguments{
\item{ground_truth}{A ground truth polygon}

\item{predictions}{prediction polygons}
}
\value{
A data frame with the crown ID, the prediction ID and the area of overlap.
}
\description{
Find area of overlap among all sets of polygons
}
