% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTreeXML.R
\name{readTreeXML}
\alias{readTreeXML}
\title{Read tree annotations for a NEON site from a set of XML files}
\usage{
readTreeXML(siteID)
}
\arguments{
\item{siteID}{NEON site abbreviation (e.g. "HARV")}
}
\value{
a dataframe of tree annotations in the format xmin, xmax, ymin, ymax
}
\description{
Read tree annotations for a NEON site from a set of XML files
}
