% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_parse.R
\name{xml_parse}
\alias{xml_parse}
\title{Parse xml tree annotations from an XML file}
\usage{
xml_parse(fil)
}
\arguments{
\item{fil}{"Character" filename of the .xml file}
}
\value{
a dataframe of tree annotations in the format xmin, xmax, ymin, ymax
}
\description{
Parse xml tree annotations from an XML file
}
