% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_to_spatial_polygons.R
\name{xml_to_spatial_polygons}
\alias{xml_to_spatial_polygons}
\title{Convert xml annotations into projected bounding boxes.
In order to plot the annotations, they need to be projected and overlayed on RGB data}
\usage{
xml_to_spatial_polygons(xml_data, raster_object)
}
\arguments{
\item{xml_data}{a dataframe returned from readTreeXML}

\item{raster_object}{a rgb raster::stack to overlay annotations}
}
\value{
SpatialPolygons object of annotations
}
\description{
Convert xml annotations into projected bounding boxes.
In order to plot the annotations, they need to be projected and overlayed on RGB data
}
