% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxes_to_spatial_polygons.R
\name{boxes_to_spatial_polygons}
\alias{boxes_to_spatial_polygons}
\title{Convert xml annotations into projected bounding boxes.
In order to plot the annotations, they need to be projected and overlayed on RGB boxes}
\usage{
boxes_to_spatial_polygons(boxes, raster_object)
}
\arguments{
\item{boxes}{a boxesframe with xmin, xmax, ymin, ymax columns. Each row is a tree bounding box}

\item{raster_object}{a rgb raster::stack to overlay annotations}
}
\value{
SpatialPolygons object of annotations
}
\description{
Convert xml annotations into projected bounding boxes.
In order to plot the annotations, they need to be projected and overlayed on RGB boxes
}
