% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_plot.R
\name{evaluate_plot}
\alias{evaluate_plot}
\title{Parse xml tree annotations from an XML file}
\usage{
evaluate_plot(ground_truth, predictions, threshold = 0.5)
}
\arguments{
\item{ground_truth}{SpatialPolygonDataFrame of ground truth polygons}

\item{predictions}{SpatialPolygonDataFrame of prediction polygons}

\item{threshold}{Intersection over Union threshold. default is 0.5}
}
\value{
recall and precision scores for the plot
}
\description{
Parse xml tree annotations from an XML file
}
